requirements.txt
colorama==0.4.6
diskcache==5.6.3
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
setuptools==80.9.0
sympy==1.14.0
torch==2.7.1
typing_extensions==4.14.1
keyboard==0.13.5
pyclog==0.1.2